// src/pages/types/TVShowEpisodes.types.ts
// Generated by https://quicktype.io

export interface TVMazeShowEpisodes {
  id: number;
  url: string;
  name: string;
  season: number;
  number: number;
  airdate: string;
  airtime: string;
  airstamp: string;
  runtime: number;
  image: Image | null;
  summary: null | string;
  _links: Links;
}
interface Links {
  self: Self;
}
interface Self {
  href: string;
}
interface Image {
  medium: string;
  original: string;
}